docs: create `Home` docs page
