+++ docs/src/pages/Home/index.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+import HomepageFeatures from '../../components/HomepageFeatures';
+import HomepageHeader from '../../components/HomepageHeader';
+
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+import Layout from '@theme/Layout';
+
+const Home: React.FC = () => {
+  const { siteConfig } = useDocusaurusContext();
+  return (
+    <Layout description={'Description will go into a meta tag in <head />'} title={`Hello from ${siteConfig.title}`}>
+      <HomepageHeader />
+      <main>
+        <HomepageFeatures />
+      </main>
+    </Layout>
+  );
+};
+
+export default Home;
